Move all editor in editor crate folder
