+++ crates/editor/Cargo.toml
@@ -6,3 +6,20 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+bevy.workspace = true
+editor_core.workspace = true
+prefab.workspace = true
+undo.workspace = true
+persistence.workspace = true
+shared.workspace = true
+
+bevy_egui = "0.23"
+egui-gizmo = "0.12"
+egui_dock = "0.8"
+egui_file = "0.11"
+
+bevy_debug_grid = "0.3"
+
+bevy_mod_picking = {version = "0.17", default-features = false, features = ["backend_raycast", "selection"]}
+bevy-inspector-egui = { version = "0.21", features = ["bevy_pbr", "highlight_changes"]}
+bevy_panorbit_camera = "0.9"
