+++ crates/editor/src/bot_menu.rs
