+++ crates/editor/src/debug_panels.rs
