+++ crates/editor/src/editor_tab.rs
