+++ crates/editor/src/hierarchy.rs
