+++ crates/editor/src/inspector/mod.rs
