+++ crates/editor/src/settings.rs
