+++ crates/editor/src/ui_registration.rs
