+++ crates/editor_core/src/lib.rs
@@ -26,6 +26,7 @@ use prefab::save::{SaveConfig, SaveState};
 use shared::*;
 use undo::{AppAutoUndo, UndoPlugin};
 
+
 pub struct EditorCore;
 
 impl Plugin for EditorCore {
