+++ crates/prefab/Cargo.toml
@@ -7,18 +7,7 @@ edition = "2021"
 
 [dependencies]
 bevy.workspace = true
-bevy_egui.workspace = true
-egui-gizmo.workspace = true
-bevy-scene-hook.workspace = true
 ron.workspace = true
 serde.workspace = true
-bevy_panorbit_camera.workspace = true
-bevy-inspector-egui.workspace = true
-pretty-type-name.workspace = true
-bevy_mod_picking.workspace = true
-bevy_debug_grid.workspace = true
-egui_dock.workspace = true
-egui_file.workspace = true
 
 shared.workspace = true
-undo.workspace = true
